--- a/pkgs/misc/sagetex/default.nix
+++ b/pkgs/misc/sagetex/default.nix
@@ -1,10 +1,11 @@
 { lib
 , stdenv
 , fetchFromGitHub
+, writeShellScript
 , texlive
 }:
 
-stdenv.mkDerivation (finalAttrs: rec {
+stdenv.mkDerivation rec {
   pname = "sagetex";
   version = "3.6.1";
 
@@ -15,8 +16,14 @@ stdenv.mkDerivation (finalAttrs: rec {
     sha256 = "sha256-OfhbXHbGI+DaDHqZCOGiSHJPHjGuT7ZqSEjKweloW38=";
   };
 
-  buildInputs = [
+  outputs = [ "tex" ];
+
+  nativeBuildInputs = [
     texlive.combined.scheme-basic
+    # multiple-outputs.sh fails if $out is not defined
+    (writeShellScript "force-tex-output.sh" ''
+      out="''${tex-}"
+    '')
   ];
 
   buildPhase = ''
@@ -29,11 +36,6 @@ stdenv.mkDerivation (finalAttrs: rec {
     cp -va *.sty *.cfg *.def "$path/"
   '';
 
-  passthru = {
-    tlType = "run";
-    pkgs = [ finalAttrs.finalPackage ];
-  };
-
   meta = with lib; {
     description = "Embed code, results of computations, and plots from Sage into LaTeX documents";
     homepage = "https://github.com/sagemath/sagetex";
@@ -41,4 +43,4 @@ stdenv.mkDerivation (finalAttrs: rec {
     maintainers = with maintainers; [ alexnortung ];
     platforms = platforms.all;
   };
-})
+}