--- a/pkgs/tools/typesetting/tex/texlive/combine.nix
+++ b/pkgs/tools/typesetting/tex/texlive/combine.nix
@@ -115,7 +115,7 @@ in (buildEnv {
 
   passthru = {
     # This is set primarily to help find-tarballs.nix to do its job
-    packages = pkgList.all;
+    packages = lib.filter lib.isDerivation pkgList.all;
     # useful for inclusion in the `fonts.packages` nixos option or for use in devshells
     fonts = "${texmfroot}/texmf-dist/fonts";
   };