@xworld21 2023-08-28
![image](https://github.com/NixOS/nixpkgs/assets/1962985/56770c8e-efd1-42a0-acb8-f7969d62c1b0)
At last! And we'll get more with `buildEnv` and the setup hooks.