@apfelkuchen6 2023-08-28
Hm, I cannot reproduce this eval performance improvement. In fact, we appear to have a recent regression:

Instantiatiating `texlive.combined.scheme-full`  takes:
this PR:  2.1s
current master: 2.0s
after [bin containers](https://github.com/NixOS/nixpkgs/commit/aa725d5a5bc3e782cf7cb872616c9c3cf5330162): 1.8s
before bin containers, 1.5s

Bisect between the bin containers and master is running.