@apfelkuchen6 2023-08-28
Ok, the eval performance regression is due to recent python changes. Nothing we can do about that:

only skipped commits left to test
possible first bad commit: [e1cafe642a5e7d265d7b1dc55e2ddda81d219377] python3.pkgs.build: move tests to passthru
possible first bad commit: [349f57fb426225483f12e082e5ab5da8137eefe8] python3.pkgs.pyproject-hooks: move tests to passthru
possible first bad commit: [0d7647889c6701d8a0e6fa784e886012011ca2d8] python3.pkgs.installer: move tests to passthru
possible first bad commit: [6c85fff302615c62bf4f632bca661bc48298b0a3] python3.pkgs.buildPythonPackage: switch to PyPA build and installer