@xworld21 2023-09-03
Rebased against new `texlive.bin.core` changes. I decided to flip the switch: there is no scenario without `texlive.pkgs`, so I made that official in the docs and applied the relevant in the rest of the tree. I also added a fix to make `texlive.combine` vaguely aware of `outputSpecified` – but that would be a new usage, so it will most likely be superseded by `withPackages`.

(I am also curious to see the *actual* evaluation report against the latest master.)